talleyrand/requirements.txt
beautifulsoup4==4.13.4
fastapi==0.116.1
langchain==0.3.27
langchain_core==0.3.72
langchain_google_genai==2.1.8
langchain_openai==0.3.28
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.36.0
